F_reentry_ AND_parameters/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
PyYAML>=6.0
jsonschema>=4.18
